% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_prop.R
\name{ci_prop}
\alias{ci_prop}
\title{Confidence interval in a pool with two candidates.}
\usage{
ci_prop(x, n, alpha = 0.05)
}
\arguments{
\item{\code{x}}{Number of votes to candidate 1.}

\item{\code{n}}{Sample size.}

\item{\code{alpha}}{Significance level. Default: \code{alpha=0.05}.}
}
\description{
Confidence interval in a pool with two candidates.
}
\examples{
ic(217, 398, .05)
ic(181, 398, .05)
ic(170, 568, .05)
}
