% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{bayes}
\alias{bayes}
\title{Probabilities in a brazilian electoral scenario.}
\usage{
bayes(p = vector(), n, M = 10^4)
}
\arguments{
\item{\code{p}}{A vector containing the proportions of votes.}

\item{\code{n}}{Sample size.}

\item{\code{M}}{Number of replications. Default: \code{10^4}.}
}
\description{
Calculates posterior probabilities via Monte Carlo.
}
\examples{
bayes(c(.4,.3,.3), 1000)
bayes(c(.3,.25,.2,.1,.05), 100)
bayes(rep(1/5,5), 500)
bayes(c(.5813972562, .3158114522, .1027912917), 50)
bayes(c(.5144202347, .3246860305, .1608937348), 100)
bayes(c(.4160925601, .3316216347, .2522858052), 500)
bayes(c(.3919345050, .3324785813, .2755869137), 10^3)
bayes(c(.3518464606, .3332479566, .3149055828), 10^4)
bayes(c(.3391808234, .3333247966, .3274943799), 10^5)
bayes(c(.3333333335, .3333333333, .3333333331), 10^20)
}
